requirements.txt
asgiref==3.4.1
Django==3.0.5
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.0.0
djongo==1.3.3
dnspython==2.1.0
mango-jwt==1.3.2
passlib==1.7.4
PyJWT==2.0.0
pymongo==3.12.1
pytz==2021.3
sqlparse==0.2.4